lesson_10/requirements.txt
selenium==4.0.0  # или актуальная версия 
pytest==7.0.0     # или актуальная версия 
allure-pytest==2.9.0  # для интеграции Allure с pytest 
